Introduce exit directive
It wasn't possible to do things that the underlying
service permits, such as specifying a result and
status when tail-calling.

To solve the issue, the SDK now exports an exit
directive that matches the underlying RPC. The
existing value/error/tail_call helpers now all use
it.

CallResult has a few convenience methods added to
help with common use cases.