Add getting started example
This example is tested. To make this work, there is a symblink from examples to
tests/examples. Also switched the test runner to pytest (a community favorite)
instead of the built-in unittest because the latter cannot ignore a directory.
As a result it would still discover the tests in examples/, failing.

To make setup easier for users, there is a new variant of the package:
`dispatch-functions[fastapi]`. It installs the package and ensures fastapi and
httpx are installed.