Remove httpx exception checks in test package
We've moved away from the built-in test server
in favor of pushing users towards using the CLI.
The CLI sets environment variables, so there's
no chance of user error which is what these messages
were designed to help with.

Eliminating the httpx dependency means we can adapt
the test package to other web frameworks/clients.